Move the Buffer support to GVRBitmapTexture Remove the GL calls from Java
